[refactor] move SafeArea to global styles.
