+++ src/app/(Authentication)/SignUp/Email/index.tsx
@@ -10,6 +10,7 @@ import {
   ButtonTextWhite,
 } from '../../../../Components/AuthButton/AuthButton';
 import AuthInput from '../../../../Components/AuthInput/AuthInput';
+import { SafeArea } from '../../../../styles/global';
 import { emailExists } from '../../../../supabase/queries/auth';
 import {
   ContentContainer,
@@ -17,9 +18,8 @@ import {
   ErrorMessageText,
   InputBoxContainer,
   InstructionContainer,
-  SafeArea,
   TitleText,
-} from '../../styles';
+} from '../../AuthFlowStyles';
 
 export default function SignUpScreen() {
   const [name, setName] = useState<string>('');
