+++ src/styles/global.ts
@@ -1,0 +1,14 @@
+import { SafeAreaView } from 'react-native';
+import styled from 'styled-components/native';
+
+import { colors } from './colors';
+
+export const SafeArea = styled(SafeAreaView)`
+  background-color: ${colors.white};
+  width: 100%;
+  min-height: 100%;
+  flex-direction: column;
+  justify-content: start;
+  align-items: center;
+  flex: 1;
+`;
